requirements.txt
Django>=6.0.1
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
psycopg[binary]>=3.3.2
dj-database-url>=2.2
python-dotenv>=1.0
drf-spectacular>=0.27

pytest>=8.0
pytest-django>=4.8